fixed favourites not loading bug

+++ app/src/main/java/com/cmpt362/nearby/classes/Post.kt
@@ -19,5 +19,5 @@ class Post(
     val iconType: IconType = IconType.NONE,
     val iconColor: Color = Color.GREY,
     val isEvent: Boolean = false,
-    val favouritesCounter: Int = 0
+    val favouritesCounter: Long = 0
 )
